backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "clerk-backend-api>=5.0.6",
    "fastapi>=0.135.2",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
